Refactor notification module
